.history/requirements_20250221215408.txt
sentence-transformers == 2.2.2
langchain
flask
pypdf
python-dotenv
pinecone[gr
langchain-pinecone
langchain-community
langchain-openai
langchain-experimental
-e .